Refactored assign, remind to reminder.js
